+++ src/settings/ProfileDetails/DeleteModals/index.js
@@ -1,0 +1,3 @@
+export * from './JobProfileDeleteModal';
+export * from './MappingProfileDeleteModal';
+