+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Eliminate identifier selection for exports with default profiles. Refs UIDEXP-413.
 * Locking and unlocking mapping profile. Refs UIDEXP-414.
 * Update modal messages related to deleting job profiles. Refs UIDEXP-417.
+* Update deleting of mapping profiles. Refs UIDEXP-418.
 
 ## [7.0.0](https://github.com/folio-org/ui-data-export/tree/v7.0.0) (2025-03-12)
 [Full Changelog](https://github.com/folio-org/ui-data-export/compare/v6.2.0...v7.0.0)